Ebook/requirements.txt
flask==2.2.2
flask-mongoengine==1.0.0
flask-restful==0.3.9
flask-bcrypt==1.0.1
flask-jwt-extended==4.4.4
flask-mail==0.9.1
flask_cors==3.0.10
minio==7.1.13
python-dotenv
pytesseract==0.3.10
Pillow==10.0.0
PyMuPDF==1.21.1
celery[mongodb]==5.2.7
redis
flower==1.2.0
requests
gTTS
pydub==0.25.1
nltk==3.8.1
Unidecode==1.3.6
langdetect==1.0.9
google-cloud-texttospeech==2.14.1
autocorrect==2.6.1
deskew==1.4.3
numpy==1.23.0
